compile-time enforcement of printf strings

+++ firmware/ext/atwinc1500/driver/source/nmspi.cpp
@@ -559,7 +559,7 @@ _FAIL_:
 		nm_bsp_sleep(1);
 		spi_cmd(CMD_RESET, 0, 0, 0, 0);
 		spi_cmd_rsp(CMD_RESET);
-		M2M_ERR("Reset and retry %d %x %x\n",retry,addr,u32data);
+		M2M_ERR("Reset and retry %d %x %x\n", retry, (unsigned int)addr, (unsigned int)u32data);
 		nm_bsp_sleep(1);
 		retry--;
 		if(retry) goto _RETRY_;
@@ -618,7 +618,7 @@ _FAIL_:
 		nm_bsp_sleep(1);
 		spi_cmd(CMD_RESET, 0, 0, 0, 0);
 		spi_cmd_rsp(CMD_RESET);
-		M2M_ERR("Reset and retry %d %x %d\n",retry,addr,size);
+		M2M_ERR("Reset and retry %d %x %d\n", retry, (unsigned int)addr, size);
 		nm_bsp_sleep(1);
 		retry--;
 		if(retry) goto _RETRY_;
